+++ src/main/java/sptech/elderly/web/dto/mensagem/MensagemInput.java
@@ -3,7 +3,7 @@ package sptech.elderly.web.dto.mensagem;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record MensagemInput(@NotNull Integer remetenteId,
-                            @NotNull Integer destinatarioId,
+public record MensagemInput(@NotNull Long remetenteId,
+                            @NotNull Long destinatarioId,
                             @NotBlank String conteudo) {
 }
